+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - unit tests for `ShareLink` component and integrate i18n translations
 - unit tests for `Participant` component and integrate i18n translations
 - unit tests for `Person` and `ProgramList` components, integrate i18n translations
+- unit tests for `TagSelectors` and `LocalTime` with i18n integration
 
 ## [0.1.0] - 2026-02-12
 - Initial release.